[py2py3] Manual changes on top futurize - WMCore/src/python/Utils
grouped imports
  * Utils/EmailAlert.py
  * Utils/IteratorTools.py

Dictionaries: replaced mydict.items() with iteritems(mydict), from future
Resources:
  * http://python-future.org/compatible_idioms.html#dictionaries
  * http://python-future.org/what_else.html#dict
  * https://stackoverflow.com/questions/21783763/in-python-2-what-are-the-performance-trade-offs-for-viewvalues-viewitems-ov
Files
  * Utils/ExtendedUnitTestCase.py
  * Utils/IteratorTools.py
  * Utils/ProcessStats.py
  * Utils/Signals.py
  * Utils/Utilities.py

Int & Long
Resources:
  * http://python-future.org/compatible_idioms.html#long-integers
  * http://python-future.org/what_else.html#int
Files:
  * Utils/MathUtils.py

+++ src/python/Utils/MathUtils.py
@@ -4,6 +4,7 @@ Module containing mathematical and physics utils
 """
 from __future__ import division, print_function
 
+from builtins import int
 from past.builtins import basestring
 from math import ceil
 
@@ -20,7 +21,7 @@ def quantize(inputVal, quanta):
         msg = "Input value has to be either int or float, not %s" % (type(inputVal))
         raise ValueError(msg)
 
-    if isinstance(quanta, (basestring, float, int)):
+    if isinstance(quanta, (basestring, int, float)):
         quanta = int(float(quanta))
     elif not isinstance(quanta, int):
         msg = "Quanta value has to be either int or float, not %s" % (type(quanta))
